Moved folder services from domain to persistence
